TicketResponseDto added; fixed some dtos and OrderController
