+++ src/main/java/cinema/model/dto/cinemahall/CinemaHallRequestDto.java
@@ -1,17 +1,8 @@
-package cinema.model.dto;
+package cinema.model.dto.cinemahall;
 
-public class CinemaHallResponseDto {
-    private Long id;
+public class CinemaHallRequestDto {
     private int capacity;
     private String description;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
 
     public int getCapacity() {
         return capacity;
