+++ src/main/java/cinema/model/dto/movie/MovieResponseDto.java
@@ -1,8 +1,17 @@
-package cinema.model.dto;
+package cinema.model.dto.movie;
 
-public class MovieRequestDto {
+public class MovieResponseDto {
+    private Long id;
     private String title;
     private String description;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
 
     public String getTitle() {
         return title;
