+++ src/main/java/cinema/model/dto/moviesession/MovieSessionResponseDto.java
@@ -1,4 +1,4 @@
-package cinema.model.dto;
+package cinema.model.dto.moviesession;
 
 public class MovieSessionResponseDto {
     private Long id;
