+++ src/main/java/cinema/model/dto/order/OrderResponseDto.java
@@ -1,4 +1,6 @@
-package cinema.model.dto;
+package cinema.model.dto.order;
+
+import cinema.model.dto.ticket.TicketResponseDto;
 
 import java.time.LocalDateTime;
 import java.util.List;
