+++ src/main/java/cinema/model/dto/shoppingcart/ShoppingCartDtoMapper.java
@@ -1,9 +1,11 @@
 
-package cinema.model.dto;
+package cinema.model.dto.shoppingcart;
 
 import cinema.model.ShoppingCart;
 import java.util.List;
 import java.util.stream.Collectors;
+
+import cinema.model.dto.ticket.TicketResponseDto;
 import org.springframework.stereotype.Component;
 
 @Component
