+++ src/main/java/cinema/model/dto/user/UserRequestDto.java
@@ -1,4 +1,4 @@
-package cinema.model.dto;
+package cinema.model.dto.user;
 
 public class UserRequestDto {
     private String email;
